+++ src/Validator/Rules/PossibleFragmentSpreads.php
@@ -37,7 +37,7 @@ class PossibleFragmentSpreads extends ValidationRule
                 }
 
                 $context->reportError(new Error(
-                    self::typeIncompatibleAnonSpreadMessage($parentType, $fragType),
+                    static::typeIncompatibleAnonSpreadMessage($parentType, $fragType),
                     [$node]
                 ));
             },
@@ -55,14 +55,14 @@ class PossibleFragmentSpreads extends ValidationRule
                 }
 
                 $context->reportError(new Error(
-                    self::typeIncompatibleSpreadMessage($fragName, $parentType, $fragType),
+                    static::typeIncompatibleSpreadMessage($fragName, $parentType, $fragType),
                     [$node]
                 ));
             },
         ];
     }
 
-    private function doTypesOverlap(Schema $schema, CompositeType $fragType, CompositeType $parentType)
+    protected function doTypesOverlap(Schema $schema, CompositeType $fragType, CompositeType $parentType)
     {
         // Checking in the order of the most frequently used scenarios:
         // Parent type === fragment type
@@ -139,7 +139,7 @@ class PossibleFragmentSpreads extends ValidationRule
         );
     }
 
-    private function getFragmentType(ValidationContext $context, $name)
+    protected function getFragmentType(ValidationContext $context, $name)
     {
         $frag = $context->getFragment($name);
         if ($frag) {